feat: allow commit hashes as actions version
Some organizations have a policy of always referring to GitHub actions by
commit hash rather than tag/branch, as tags/branches can be moved without
review.

Since we don't want to be pressured into releasing a new version of
sbt-github-actions each time an action publishes a release, this changes
`githubWorkflowCheck` to ignore hashes. This might make the 'diff'
slightly confusing because the 'actual' will be the actual file with
the hashes replaced, but that seems preferable over highlighting
the hashes in the diff.