Allow any action to be referenced using commit hash
Support for 'tolerating' commit hashes as refs without causing
`githubWorkflowCheck` to fail was introduced in #207. However, this does
not permit adding additional steps or jobs which uses actions not listed
in `Action.all`, since they will not be taken into consideration when
replacing refs in the comparison.

This CL replaces the `Action.all` value to instead traverse the
generated workflow structure to find all used actions.